+++ beam-core/src/test/java/org/esa/beam/framework/dataio/ProductFlipperTest.java
@@ -67,12 +67,12 @@ public class ProductFlipperTest {
     public void testTiePointFlipping_Horizontal() throws IOException {
         final Product flipped = ProductFlipper.createFlippedProduct(product, ProductFlipper.FLIP_HORIZONTAL, "h", "h");
         final TiePointGrid flippedT1 = flipped.getTiePointGrid("t1");
-        final float[] expectedT1 = {
+        final double[] expectedT1 = {
                 0.4f, 0.3f, 0.6f,
                 0.4f, 0.9f, 0.8f,
                 0.4f, 0.2f, 0.3f
         };
-        assertTrue(Arrays.equals(expectedT1, (float[]) flippedT1.getDataElems()));
+        assertTrue(Arrays.equals(expectedT1, (double[]) flippedT1.getDataElems()));
 
         assertEquals(4, flippedT1.getOffsetX(), 1.0e-6);
         assertEquals(0, flippedT1.getOffsetY(), 1.0e-6);
@@ -99,12 +99,12 @@ public class ProductFlipperTest {
     public void testTiePointFlipping_Vertical() throws IOException {
         final Product flipped = ProductFlipper.createFlippedProduct(product, ProductFlipper.FLIP_VERTICAL, "v", "v");
         final TiePointGrid flippedT1 = flipped.getTiePointGrid("t1");
-        final float[] expectedT1 = {
+        final double[] expectedT1 = {
                 0.3f, 0.2f, 0.4f,
                 0.8f, 0.9f, 0.4f,
                 0.6f, 0.3f, 0.4f
         };
-        assertTrue(Arrays.equals(expectedT1, (float[]) flippedT1.getDataElems()));
+        assertTrue(Arrays.equals(expectedT1, (double[]) flippedT1.getDataElems()));
 
         assertEquals(0, flippedT1.getOffsetX(), 1.0e-6);
         assertEquals(6, flippedT1.getOffsetY(), 1.0e-6);
@@ -131,12 +131,12 @@ public class ProductFlipperTest {
     public void testTiePointFlipping_Both() throws IOException {
         final Product flipped = ProductFlipper.createFlippedProduct(product, ProductFlipper.FLIP_BOTH, "b", "b");
         final TiePointGrid flippedT1 = flipped.getTiePointGrid("t1");
-        final float[] expectedT1 = {
+        final double[] expectedT1 = {
                 0.4f, 0.2f, 0.3f,
                 0.4f, 0.9f, 0.8f,
                 0.4f, 0.3f, 0.6f
         };
-        assertTrue(Arrays.equals(expectedT1, (float[]) flippedT1.getDataElems()));
+        assertTrue(Arrays.equals(expectedT1, (double[]) flippedT1.getDataElems()));
 
         assertEquals(4, flippedT1.getOffsetX(), 1.0e-6);
         assertEquals(6, flippedT1.getOffsetY(), 1.0e-6);
